Move detect_img() and all commmand line parsing function to yolo_video.py.  Added -i | --image flag for image detection mode.  Default is video detection mode.
